refactor(orgs): retire REST org service, migrate all callers to GQL
- Move organizationKeys into gql-service; delete organization.service.ts
  and organization.queries.ts
- Add useOrganizations / useOrganization TanStack hooks backed by GQL
  service (same signatures, no call-site changes needed)
- Migrate 6 SSR loaders and 3 CSR mutation hooks to GQL equivalents
- Fix org adapter: coerce GQL null → undefined for namespace
  (z.string().optional() rejects null; GraphQL returns it for
  cluster-scoped resources)
- Fix create/update mutations: add apiVersion+kind to request body
- Remove unused REST session hooks from user.queries.ts
